Add provider trait for HasConnectionOpenInitEvent
