+++ crates/cosmos/cosmos-chain-components/src/impls/events.rs
@@ -2,15 +2,21 @@ use alloc::sync::Arc;
 
 use hermes_relayer_components::chain::traits::types::create_client::ProvideCreateClientEvent;
 use hermes_relayer_components::chain::traits::types::ibc::HasIbcChainTypes;
-use hermes_relayer_components::chain::traits::types::ibc_events::connection::ProvideConnectionOpenInitEvent;
-use ibc_relayer::event::connection_open_ack_try_from_abci_event;
+use hermes_relayer_components::chain::traits::types::ibc_events::connection::{
+    ProvideConnectionOpenInitEvent, ProvideConnectionOpenTryEvent,
+};
+use ibc_relayer::event::{
+    connection_open_ack_try_from_abci_event, connection_open_try_try_from_abci_event,
+};
 use ibc_relayer_types::core::ics02_client::events::CLIENT_ID_ATTRIBUTE_KEY;
 use ibc_relayer_types::core::ics24_host::identifier::{ClientId, ConnectionId};
 use ibc_relayer_types::events::IbcEventType;
 use tendermint::abci::Event as AbciEvent;
 
 use crate::types::events::client::CosmosCreateClientEvent;
-use crate::types::events::connection::CosmosConnectionOpenInitEvent;
+use crate::types::events::connection::{
+    CosmosConnectionOpenInitEvent, CosmosConnectionOpenTryEvent,
+};
 
 pub struct ProvideCosmosEvents;
 
@@ -70,6 +76,35 @@ where
 
     fn connection_open_init_event_connection_id(
         event: &CosmosConnectionOpenInitEvent,
+    ) -> &ConnectionId {
+        &event.connection_id
+    }
+}
+
+impl<Chain, Counterparty> ProvideConnectionOpenTryEvent<Chain, Counterparty> for ProvideCosmosEvents
+where
+    Chain: HasIbcChainTypes<Counterparty, Event = Arc<AbciEvent>, ConnectionId = ConnectionId>,
+{
+    type ConnectionOpenTryEvent = CosmosConnectionOpenTryEvent;
+
+    fn try_extract_connection_open_try_event(
+        event: Arc<AbciEvent>,
+    ) -> Option<CosmosConnectionOpenTryEvent> {
+        let event_type = event.kind.parse().ok()?;
+
+        if let IbcEventType::OpenTryConnection = event_type {
+            let open_try_event = connection_open_try_try_from_abci_event(&event).ok()?;
+
+            let connection_id = open_try_event.connection_id()?.clone();
+
+            Some(CosmosConnectionOpenTryEvent { connection_id })
+        } else {
+            None
+        }
+    }
+
+    fn connection_open_try_event_connection_id(
+        event: &CosmosConnectionOpenTryEvent,
     ) -> &ConnectionId {
         &event.connection_id
     }
